Create Часть 1 глава1.1
The first time I woke up by myself. Either the sleeping pills turned out to be unusable, or Max broke down.

I opened my eyes and saw nothing. After all, he simply could not see anything in this pitch darkness. And despite the fact that the sun was already shining in full force, the night continued in my house.

The darkness lasted only a moment, but this time was enough for me to feel an inexplicable languor in my chest. He was born with a desire that only teenagers can experience, at those moments when the unknown in this world beckons and calls to be known.

I did not know the name of this emotion, because in our world there is nothing unknown. We have nothing more to strive for. We can find the answer to any question that interests us in an instant. And just as quickly forget it. Why remember him? After all, now there is a powerful information storage system, which is always at hand, or rather before our eyes.